dialog could use true in isVisible and change file
