+++ src/feature/modal-reply/style.module.scss
@@ -1,0 +1,77 @@
+.modal {
+  &__post {
+    padding: 0.9375rem;
+    display: grid;
+    grid-template-columns: auto 1fr;
+    gap: 0.625rem;
+    min-height: 15.3125rem;
+    font-size: 18px;
+  }
+
+  &__reply {
+    padding: 1.125rem 0.9375rem 0.625rem;
+    gap: 0.625rem;
+    font-size: 15px;
+    display: grid;
+    grid-template-columns: auto 1fr;
+  }
+
+  &__content {
+    word-wrap: break-word;
+    color: #1c1c1c;
+    & p:first-child {
+      margin-block-end: 0.3125rem;
+      & span {
+        color: #657786;
+      }
+    }
+
+    & p:last-child {
+      margin-block-start: 1.1875rem;
+      font-size: 13px;
+      color: #657786;
+
+      & a {
+        color: #ff6600;
+      }
+    }
+  }
+}
+
+.avatar {
+  width: 3.125rem;
+  & > * {
+    cursor: pointer;
+  }
+}
+
+.form {
+  display: grid;
+  grid-template-rows: 1fr auto;
+  gap: 0.625rem;
+  padding-block-start: 0.75rem;
+
+  &__btn {
+    margin-inline-start: auto;
+    & span {
+      color: white;
+      font-weight: 500;
+    }
+  }
+}
+
+.textarea {
+  border: none;
+  resize: none;
+  font-size: inherit;
+  &:focus-visible {
+    outline: none;
+  }
+}
+
+.decoration {
+  height: calc(100% - 1.5626rem - 0.625rem);
+  background-color: #ccd6dd;
+  width: 1px;
+  translate: 1.5rem;
+}
